+++ src/domain/handlers/interfaces/imovie_handler.py
@@ -1,16 +1,7 @@
-from abc import ABC, abstractmethod
+from abc import ABC
 
-from telegram import Update
-from telegram.ext import CallbackContext
+from src.domain.handlers.interfaces.imedia_handler import IMediaHandler
 
 
-class IMovieHandler(ABC):
-    @abstractmethod
-    def get_folders(self, update: Update, context: CallbackContext):
-        """Get folders or set default folder"""
-    @abstractmethod
-    def get_quality_profiles(self, update: Update, context: CallbackContext):
-        """Get quality profiles or set default quality profile"""
-    @abstractmethod
-    def add_to_library(self, update: Update, context: CallbackContext) -> None | int:
-        """Add the selected media to the library"""
+class IMovieHandler(IMediaHandler, ABC):
+    """Movie handler interface"""
